@@ -59,7 +59,7 @@
   ],
   "dependencies": {
     "argparse": "^2.0.1",
-    "dotenv": "^10.0.0",
+    "dotenv": "^16.0.0",
     "iconv-lite": "^0.6.3",
     "js-base64": "^3.7.2",
     "tmp": "^0.2.1",
@@ -71,28 +71,28 @@
   },
   "devDependencies": {
     "@types/argparse": "^2.0.10",
-    "@types/chai": "^4.2.22",
+    "@types/chai": "^4.3.0",
     "@types/glob": "^7.2.0",
-    "@types/mocha": "^9.0.0",
-    "@types/node": "^16.11.6",
+    "@types/mocha": "^9.1.0",
+    "@types/node": "^17.0.18",
     "@types/semver": "^7.3.9",
-    "@types/tmp": "^0.2.2",
-    "@types/vscode": "^1.23.0",
+    "@types/tmp": "^0.2.3",
+    "@types/vscode": "^1.59.0",
     "@types/xml2js": "^0.4.9",
-    "chai": "^4.3.4",
+    "chai": "^4.3.6",
     "copyfiles": "2.4.1",
     "cross-env": "^7.0.3",
     "glob": "^7.2.0",
-    "mocha": "^9.1.3",
+    "mocha": "^9.2.1",
     "rimraf": "^3.0.2",
     "tslint": "^6.1.3",
-    "typescript": "^4.4.4",
-    "vsce": "^1.100.2",
+    "typescript": "^4.5.5",
+    "vsce": "^1.103.1",
     "vscode-test": "^1.6.1"
   },
   "main": "out/src/main.js",
   "engines": {
-    "vscode": "^1.23.0"
+    "vscode": "^1.59.0"
   },
   "activationEvents": [
     "onLanguage:python",
