@@ -1,18 +1,16 @@
 import { exec } from 'child_process';
-import util = require('util');
-export const execPromise = util.promisify(exec);
+import * as util from 'util';
 import { getPythonExecutable } from '../utils/testConfiguration';
-import { gte } from 'semver';
+import { gte, lt } from 'semver';
 
-
+export const execPromise = util.promisify(exec);
 export async function isTestplanPrerequisiteMet(): Promise<boolean> {
     const command = getPythonExecutable() + ' --version';
     const result = await execPromise(command);
-    if (!result.stderr && result.stdout)
-    {
+    if (!result.stderr && result.stdout) {
         // stdout is "Python <major>.<minor>.<patch>"
         const version = result.stdout.split(' ')[1];
-        return gte(version, '3.7.0');
+        return gte(version, '3.7.0') && lt(version, '3.10.0');
     }
     return false;
 }
